Update wattwise.py
Problem: 
The check for cheapest / expensive hours is only done for the day the forecast is published at 2 p.m. for the same day as it compares now.date and forecast.date. 
When there are cheap hours e.g. 2-4 p.m. this very day but even cheaper hours from e.g. 11-14h the next day, the script will only pull the cheap hours from 2-4 p.m. but ignore the cheap hours from tomorrow, even if the price is lower the next day.

Solution: 
When the date changes at 0:00 the script will redo the check for cheapest hours on the new day. 

changes have been made for expensive hours respectively.